feat: move smart values in a dedicated module and fix instantiation
