+++ mxops/execution/smart_values/__init__.py
@@ -1,0 +1,47 @@
+"""
+author: Etienne Wallet
+
+This subpackage contains all the smart values that can be used with MxOps
+"""
+
+from mxops.execution.smart_values.base import SmartValue
+from mxops.execution.smart_values.custom import (
+    SmartOnChainTokenTransfer,
+    SmartOnChainTokenTransfers,
+    SmartResultsSaveKeys,
+)
+from mxops.execution.smart_values.mx_sdk import (
+    SmartAddress,
+    SmartBech32,
+    SmartToken,
+    SmartTokenTransfer,
+    SmartTokenTransfers,
+)
+from mxops.execution.smart_values.native import (
+    SmartInt,
+    SmartFloat,
+    SmartBool,
+    SmartStr,
+    SmartDict,
+    SmartList,
+    SmartPath,
+)
+
+__all__ = [
+    "SmartAddress",
+    "SmartBech32",
+    "SmartBool",
+    "SmartDict",
+    "SmartFloat",
+    "SmartInt",
+    "SmartList",
+    "SmartOnChainTokenTransfer",
+    "SmartOnChainTokenTransfers",
+    "SmartPath",
+    "SmartResultsSaveKeys",
+    "SmartStr",
+    "SmartToken",
+    "SmartTokenTransfer",
+    "SmartTokenTransfers",
+    "SmartValue",
+]
