+++ mxops/execution/steps/base.py
@@ -14,11 +14,8 @@ from mxops.execution.account import AccountsManager
 from mxops.execution.checks.factory import SmartChecks
 from mxops.execution.checks import SuccessCheck
 from mxops.execution.network import send, send_and_wait_for_result
-from mxops.execution.smart_values import (
-    SmartStr,
-    SmartValue,
-    extract_first_smart_value_class,
-)
+from mxops.execution.smart_values import SmartStr, SmartValue
+from mxops.execution.smart_values.factory import extract_first_smart_value_class
 from mxops.utils.logger import get_logger
 from mxops.utils.msc import get_tx_link
